+++ frontend/package.json
@@ -40,10 +40,9 @@
     "js-cookie": "^3.0.5",
     "leaflet": "^1.9.4",
     "lottie-react": "^2.4.0",
-    "lucide-react": "^0.469.0",
     "moment": "^2.30.1",
     "moment-timezone": "^0.5.45",
-    "next": "^14.0.4",
+    "next": "14.0.4",
     "primereact": "^10.6.3",
     "react": "^18.2.0",
     "react-apexcharts": "^1.4.1",
