--- frontend/src/app/service-restriction/page.tsx
@@ -1,15 +1,0 @@
-
-import ServiceRestrictionLayout from '@/components/JuidcoHrms/pages/Error/ServiceRestrictionLayout'
-import React from 'react'
-
-const page = () => {
-
-  
-  return (
-    <div>
-      <ServiceRestrictionLayout/>
-    </div>
-  )
-}
-
-export default page
